+++ bindings/python/pyiree/compiler/compiler.cc
@@ -26,8 +26,9 @@
 #include "iree/compiler/Dialect/VM/Target/init_targets.h"
 #include "iree/compiler/Dialect/VM/Transforms/Passes.h"
 #include "iree/tools/init_compiler_modules.h"
-#include "iree/tools/init_dialects.h"
+#include "iree/tools/init_iree_dialects.h"
 #include "iree/tools/init_iree_passes.h"
+#include "iree/tools/init_mlir_dialects.h"
 #include "iree/tools/init_mlir_passes.h"
 #include "iree/tools/init_targets.h"
 #include "llvm/Support/CommandLine.h"
