+++ iree/samples/custom_modules/dialect/custom_opt.cc
@@ -23,8 +23,9 @@
 #include "iree/compiler/Dialect/HAL/Conversion/Passes.h"
 #include "iree/samples/custom_modules/dialect/init_dialect.h"
 #include "iree/tools/init_compiler_modules.h"
-#include "iree/tools/init_dialects.h"
+#include "iree/tools/init_iree_dialects.h"
 #include "iree/tools/init_iree_passes.h"
+#include "iree/tools/init_mlir_dialects.h"
 #include "iree/tools/init_mlir_passes.h"
 #include "iree/tools/init_targets.h"
 #include "iree/tools/init_xla_dialects.h"
