+++ iree/tools/init_mlir_dialects.h
@@ -17,16 +17,9 @@
 //
 // Based on MLIR's InitAllDialects but without dialects we don't care about.
 
-#ifndef IREE_TOOLS_INIT_DIALECTS_H_
-#define IREE_TOOLS_INIT_DIALECTS_H_
+#ifndef IREE_TOOLS_INIT_MLIR_DIALECTS_H_
+#define IREE_TOOLS_INIT_MLIR_DIALECTS_H_
 
-#include "iree/compiler/Dialect/Flow/IR/FlowDialect.h"
-#include "iree/compiler/Dialect/HAL/IR/HALDialect.h"
-#include "iree/compiler/Dialect/IREE/IR/IREEDialect.h"
-#include "iree/compiler/Dialect/Shape/IR/ShapeDialect.h"
-#include "iree/compiler/Dialect/VM/IR/VMDialect.h"
-#include "iree/compiler/Dialect/VMLA/IR/VMLADialect.h"
-#include "iree/compiler/Dialect/Vulkan/IR/VulkanDialect.h"
 #include "mlir/Dialect/Affine/IR/AffineOps.h"
 #include "mlir/Dialect/GPU/GPUDialect.h"
 #include "mlir/Dialect/LLVMIR/LLVMDialect.h"
@@ -63,25 +56,4 @@ inline void registerMlirDialects() {
 }
 }  // namespace mlir
 
-namespace mlir {
-namespace iree_compiler {
-
-// This function should be called before creating any MLIRContext if one expect
-// all the possible dialects to be made available to the context automatically.
-inline void registerIreeDialects() {
-  static bool init_once = []() {
-    registerDialect<IREE::Flow::FlowDialect>();
-    registerDialect<IREE::HAL::HALDialect>();
-    registerDialect<ShapeDialect>();
-    registerDialect<IREEDialect>();
-    registerDialect<IREE::VM::VMDialect>();
-    registerDialect<IREE::VMLA::VMLADialect>();
-    registerDialect<IREE::Vulkan::VulkanDialect>();
-    return true;
-  }();
-  (void)init_once;
-}
-}  // namespace iree_compiler
-}  // namespace mlir
-
-#endif  // IREE_TOOLS_INIT_DIALECTS_H_
+#endif  // IREE_TOOLS_INIT_MLIR_DIALECTS_H_
