+++ iree/tools/opt_main.cc
@@ -21,8 +21,9 @@
 #include "iree/compiler/Conversion/init_conversions.h"
 #include "iree/compiler/Dialect/HAL/Conversion/Passes.h"
 #include "iree/tools/init_compiler_modules.h"
-#include "iree/tools/init_dialects.h"
+#include "iree/tools/init_iree_dialects.h"
 #include "iree/tools/init_iree_passes.h"
+#include "iree/tools/init_mlir_dialects.h"
 #include "iree/tools/init_mlir_passes.h"
 #include "iree/tools/init_targets.h"
 #include "iree/tools/init_xla_dialects.h"
