+++ iree/tools/run_mlir_main.cc
@@ -60,7 +60,8 @@
 #include "iree/hal/api.h"
 #include "iree/modules/hal/hal_module.h"
 #include "iree/tools/init_compiler_modules.h"
-#include "iree/tools/init_dialects.h"
+#include "iree/tools/init_iree_dialects.h"
+#include "iree/tools/init_mlir_dialects.h"
 #include "iree/tools/init_targets.h"
 #include "iree/tools/vm_util.h"
 #include "iree/vm/api.h"
